requirements.txt
# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector store
supabase>=2.0.0
pgvector>=0.2.0

# Processing
langchain>=0.1.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
